# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wlanpi-core",
...
    "version": "2.1.10"
  },
  "paths": {
    "/api/v1/auth/token": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Issue JWT bearer token",
        "description": "Issue a JWT for remote clients.\n\nSend the returned `access_token` as `Authorization: Bearer <token>` on all\nsubsequent API calls until expiry (default 7 days) or `DELETE /auth/token`.",
        "operationId": "generate_token_api_v1_auth_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token / HMAC signature",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "412": {
            "description": "device_id missing from request body"
          },
          "500": {
            "description": "Token generation failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "authentication"
        ],
        "summary": "Revoke current JWT",
        "description": "Revoke the bearer token sent in the `Authorization` header.\n\nThe request body must include the same `device_id` used when the token was issued.",
        "operationId": "revoke_token_api_v1_auth_token_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenRevokeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token / HMAC signature",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "412": {
            "description": "device_id missing from request body"
          },
          "500": {
            "description": "Revocation failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable error or status message"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "MessageResponse"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "JWT bearer token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer",
            "examples": [
              "bearer"
            ]
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "Token"
      },
      "TokenRequest": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          }
        },
        "type": "object",
        "required": [
          "device_id"
        ],
        "title": "TokenRequest"
      },
      "TokenRevokeResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "success | info (already revoked) | warning (not found)",
            "examples": [
              "success",
              "info",
              "warning"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "Token revoked"
            ]
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id",
            "description": "Present when a token record was matched"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "TokenRevokeResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}